指令推送逻辑修改

+++ danmu_collector_server/src/main/java/cn/partytime/scheduler/DanmuServerScheduler.java
@@ -95,19 +95,19 @@ public class DanmuServerScheduler {
         if(ListUtils.checkListIsNotNull(addressIdList)){
             for(String addressId:addressIdList){
                 Map<String,Object>  stringObjectMap = clientCacheService.getFirstCommandFromCache(addressId);
-                String message = JSON.toJSONString(stringObjectMap);
-                //{"type":"command","data":{"clientType":"0","partyTime":1501830393838,"type":"partyStatus","partyId":"59841cedf0b0a1a490ea343c","status":2,"movieTime":1501833182944}}
-                //Map<String, Object> dataObject = (Map<String, Object>) JSON.parse(String.valueOf(stringObjectMap.get("data")));
-                int count = clientCacheService.findTempCommandCount(addressId);
-                logger.info("场地:{},重复发送指令次数:{}",addressId,count);
-                if(count<=5){
-                    clientCacheService.addTempCommandCount(addressId);
-                }else{
-                    clientCacheService.removeTempCommandCount(addressId);
-                    clientCacheService.removeFirstCommandFromCache(addressId);
+                if(stringObjectMap!=null){
+                    String message = JSON.toJSONString(stringObjectMap);
+                    int count = clientCacheService.findTempCommandCount(addressId);
+                    logger.info("场地:{},重复发送指令次数:{}",addressId,count);
+                    if(count<=5){
+                        clientCacheService.addTempCommandCount(addressId);
+                    }else{
+                        clientCacheService.removeTempCommandCount(addressId);
+                        clientCacheService.removeFirstCommandFromCache(addressId);
+                    }
+                    logger.info("向客户重新发送的命令是:{}",message);
+                    danmuSendService.pubDanmuToAllScreenClient(addressId,message);
                 }
-                logger.info("向客户重新发送的命令是:{}",message);
-                danmuSendService.pubDanmuToAllScreenClient(addressId,message);
             }
         }
     }
